# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NCDMB API Documentation",
        "description": "Comprehensive API documentation for the NCDMB Document Management System. This API provides endpoints for document management, workflow automation, procurement, legal processes, and more.",
        "contact": {
            "email": "support@ncdmb.gov.ng"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/procurement/bid-invitations": {
            "get": {
                "tags": [
                    "Procurement"
                ],
                "summary": "List all bid invitations",
                "operationId": "34a6161eb8ac8f1001e72d6f652d035b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Create a new bid invitation",
                "operationId": "de044563f6f93c28feb985b1a06e2430",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectBidInvitationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bid invitation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/procurement/bid-invitations/{id}": {
            "get": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Get a specific bid invitation",
                "operationId": "c8d1d589080377d2f0ad8b066b11df3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Update a bid invitation",
                "operationId": "20cc9d07933f398cbb6242602a4c18bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectBidInvitationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bid invitation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Delete a bid invitation",
                "operationId": "6a056b2de2e99d596dee133827610c01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bid invitation deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/procurement/bid-invitations/{id}/publish": {
            "post": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Publish a bid invitation",
                "operationId": "0be88d3b72eeaec49ad45518bc04a1ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bid invitation published successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/procurement/bid-invitations/{id}/close": {
            "post": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Close a bid invitation",
                "operationId": "ceff434959f256190e61ec4418d7e650",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bid invitation closed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Success": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    }
                },
                "type": "object"
            },
...
            "ProjectBidInvitationRequest": {
                "required": [
                    "project_id",
                    "title",
                    "submission_deadline",
                    "opening_date"
                ],
                "properties": {
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 500,
                        "example": "Infrastructure Development Bid"
                    },
                    "description": {
                        "type": "string",
                        "example": "Bid invitation description",
                        "nullable": true
                    },
                    "technical_specifications": {
                        "type": "string",
                        "example": "Technical requirements",
                        "nullable": true
                    },
                    "scope_of_work": {
                        "type": "string",
                        "example": "Scope details",
                        "nullable": true
                    },
                    "deliverables": {
                        "type": "string",
                        "example": "Expected deliverables",
                        "nullable": true
                    },
                    "terms_and_conditions": {
                        "type": "string",
                        "example": "Terms and conditions",
                        "nullable": true
                    },
                    "required_documents": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "tax_clearance",
                            "company_registration"
                        ],
                        "nullable": true
                    },
                    "eligibility_criteria": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "5_years_experience",
                            "certified_contractor"
                        ],
                        "nullable": true
                    },
                    "bid_security_required": {
                        "type": "boolean",
                        "example": true
                    },
                    "bid_security_amount": {
                        "type": "number",
                        "minimum": 0,
                        "example": 50000,
                        "nullable": true
                    },
                    "bid_security_validity_days": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 90,
                        "nullable": true
                    },
                    "estimated_contract_value": {
                        "type": "number",
                        "minimum": 0,
                        "example": 1000000,
                        "nullable": true
                    },
                    "advertisement_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15",
                        "nullable": true
                    },
                    "pre_bid_meeting_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-20",
                        "nullable": true
                    },
                    "pre_bid_meeting_location": {
                        "type": "string",
                        "maxLength": 500,
                        "example": "Conference Room A",
                        "nullable": true
                    },
                    "submission_deadline": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-15T17:00:00Z"
                    },
                    "bid_validity_days": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 120,
                        "nullable": true
                    },
                    "opening_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-16T10:00:00Z"
                    },
                    "opening_location": {
                        "type": "string",
                        "maxLength": 500,
                        "example": "Main Hall",
                        "nullable": true
                    },
                    "evaluation_criteria": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "technical_weight": {
                        "type": "number",
                        "maximum": 100,
                        "minimum": 0,
                        "example": 60,
                        "nullable": true
                    },
                    "financial_weight": {
                        "type": "number",
                        "maximum": 100,
                        "minimum": 0,
                        "example": 40,
                        "nullable": true
                    },
                    "published_newspapers": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "published_bpp_portal": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "closed",
                            "cancelled"
                        ],
                        "example": "draft",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token Authentication. Enter token in format: Bearer <token>",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}